feat: move get_max_query_range fn
